+++ src/mo_config/args.ml
@@ -90,16 +90,13 @@ let persistent_actors_args = [
   (* default stability *)
   "--default-persistent-actors",
   Arg.Unit (fun () -> Flags.actors := Flags.DefaultPersistentActors),
-  _UNDOCUMENTED_
-    " declare every actor (class) as implicitly `persistent`, defaulting actor fields to `stable` (default is --require-persistent-actors). The `persistent` keyword is now optional and redundant.";
+  " declare every actor (class) as implicitly `persistent`, defaulting actor fields to `stable` (default). The `persistent` keyword is now optional and redundant.";
 
   "--require-persistent-actors",
   Arg.Unit (fun () -> Flags.actors := Flags.RequirePersistentActors),
-  _UNDOCUMENTED_
-    " requires all actors to be declared persistent, defaulting actor fields to `transient` (default). Emit diagnostics to help migrate from non-persistent to `persistent` actors.";
+  " require all actors to be declared `persistent`, defaulting actor fields to `transient`. Emit diagnostics to help migrate from non-persistent to `persistent` actors. Use this to restore the pre-v2 default.";
 
   "--legacy-actors",
   Arg.Unit (fun () -> Flags.actors := Flags.LegacyActors),
-  _UNDOCUMENTED_
-    " in non-`persistent` actors, silently default actor fields to `transient` (legacy behaviour)";
+  " in non-`persistent` actors, silently default actor fields to `transient` (legacy behaviour)";
 ]